Reorganize store folder to be more intuitive
