+++ src/store/index.ts
@@ -3,9 +3,9 @@ import Vuex, { StoreOptions } from "vuex";
 import { vuexfireMutations } from "vuexfire";
 import collections from "@/network/collections";
 import { RootState } from "./types";
-import createCollectionModule from "./collection";
-import { actions as inventoryActions } from "./collection/modules/inventory";
-import { actions as archiveActions } from "./collection/modules/archive";
+import createCollectionModule from "./modules/collection";
+import { actions as inventoryActions } from "./modules/inventory";
+import { actions as archiveActions } from "./modules/archive";
 
 Vue.use(Vuex);
 
@@ -15,6 +15,7 @@ const store: StoreOptions<RootState> = {
   modules: {
     inventory: createCollectionModule(collections.INVENTORY, inventoryActions),
     archive: createCollectionModule(collections.ARCHIVE, archiveActions),
+    "run-preview": createCollectionModule(collections.RUNS),
   },
   strict: process.env.NODE_ENV !== "production",
 };
