+++ src/store/modules/collection/index.ts
@@ -5,7 +5,7 @@ import getters from "./getters";
 import actions from "./actions";
 import mutations from "./mutations";
 import { CollectionState } from "./types";
-import { RootState } from "../types";
+import { RootState } from "../../types";
 
 export const state: CollectionState = {
   collection: null,
